requirements.txt
pulumi>=0.16.4
pulumi_aws>=0.16.2
pulumi_kubernetes
pyyaml
